This class handle all access and management of all the particles resulting from the particles analysis. Particles are stored in a list object and methods are available for access and modification of particles and their properties.
One important note in for the architecture of the new ParticleAnalyser or ParticleManager is that particle objects are never truly removed or deleted from the manager but their visibility, through the Particles visible attribute, is modified in order to "remove" them. This is therefore non-destructive meaning that particles can always be retrieved by setting there visibility to true. Moreover, setting of hiding particles requires the mode of hiding to be specified (ie. manual delete, touching edge, size etc.). This thereby has implications and provides functionality for excluding particles by edge and size and always allows them to be retrieved.