Build: start again in new Eclipse project
Idea is to work on BoneJ1 in one project with Ant builder, jar
dependencies etc set up. In another project, work on the bonej2 branch
of the bonej Git repository, with Maven handling dependencies. This way
the code tree can stay in one piece but the two build styles are
separated.